Renamed and move rubocop rule
