+++ lib/rubocop/cop/github/accessibility/image_has_alt.rb
